Add ‘no switchport’ to switch_freeform policy
Adding ‘no switchport’ to all interfaces so that the playbooks succeed if they are run prior to user issuing Recalculate and Deploy after main switch configuration tasks.

If the user performs tasks in the following order, the mac address playbooks will fail because the interfaces are not yet routed ports.

1. Add switches
2. Run mac address playbooks
3. Recalculate and Deploy

The addition of ‘no switchport’ fixes this case.